enterprise_bot/requirements.txt
anthropic>=0.18.0
faiss-cpu>=1.7.4
fastapi>=0.109.0
hdbscan>=0.8.40
httpx>=0.27.0
ijson>=3.2.0
joblib>=1.3.0
numpy>=1.24.0
openai>=1.12.0
pydantic>=2.5.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyYAML>=6.0
rank-bm25>=0.2.2
requests>=2.31.0
river>=0.21.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
uvicorn[standard]>=0.27.0
websockets>=12.0
supabase>=2.0.0 